stactools_pipelines/pipelines/sentinel1/requirements.txt
aws-lambda-powertools
jmespath
rasterio
stactools-sentinel1
s3fs
boto3
